// src/api/models/search-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SearchResult
 */
export interface SearchResult {
    /**
     * 검색 결과 타입
     * @type {string}
     * @memberof SearchResult
     */
    'type': SearchResultTypeEnum;
    /**
     * ID (Spotify ID)
     * @type {string}
     * @memberof SearchResult
     */
    'id': string;
    /**
     * 이미지 URL
     * @type {string}
     * @memberof SearchResult
     */
    'imageUrl': string;
    /**
     * 제목 (아티스트명, 앨범명, 트랙명)
     * @type {string}
     * @memberof SearchResult
     */
    'title': string;
    /**
     * 아티스트명 (앨범/트랙인 경우)
     * @type {string}
     * @memberof SearchResult
     */
    'artist'?: string;
    /**
     * 발매 연도 (앨범/트랙인 경우)
     * @type {string}
     * @memberof SearchResult
     */
    'releaseYear'?: string;
    /**
     * 인기도 (0-100)
     * @type {number}
     * @memberof SearchResult
     */
    'popularity'?: number;
    /**
     * 평균 평점 (로그인 시 사용자 평점 포함)
     * @type {number}
     * @memberof SearchResult
     */
    'rating'?: number;
    /**
     * 현재 사용자가 평가했는지 여부 (로그인 시에만)
     * @type {boolean}
     * @memberof SearchResult
     */
    'isRated'?: boolean;
}

export const SearchResultTypeEnum = {
    Artist: 'artist',
    Album: 'album',
    Track: 'track'
} as const;

export type SearchResultTypeEnum = typeof SearchResultTypeEnum[keyof typeof SearchResultTypeEnum];