// src/api/models/create-collection-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CollectionItemInput } from './collection-item-input';

/**
 * 
 * @export
 * @interface CreateCollectionRequest
 */
export interface CreateCollectionRequest {
    /**
     * 컬렉션 제목
     * @type {string}
     * @memberof CreateCollectionRequest
     */
    'title': string;
    /**
     * 컬렉션 소개글
     * @type {string}
     * @memberof CreateCollectionRequest
     */
    'description': string;
    /**
     * 공개 여부 (public: 공개, private: 비공개)
     * @type {string}
     * @memberof CreateCollectionRequest
     */
    'isPublic': CreateCollectionRequestIsPublicEnum;
    /**
     * 컬렉션 커버 이미지 URL
     * @type {string}
     * @memberof CreateCollectionRequest
     */
    'coverImageUrl': string;
    /**
     * 초기 아이템 목록 (중복 제거됨)
     * @type {Array<CollectionItemInput>}
     * @memberof CreateCollectionRequest
     */
    'items'?: Array<CollectionItemInput>;
    /**
     * 컬렉션 태그 목록
     * @type {Array<string>}
     * @memberof CreateCollectionRequest
     */
    'tags'?: Array<string>;
}

export const CreateCollectionRequestIsPublicEnum = {
    Public: 'public',
    Private: 'private'
} as const;

export type CreateCollectionRequestIsPublicEnum = typeof CreateCollectionRequestIsPublicEnum[keyof typeof CreateCollectionRequestIsPublicEnum];